+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+* Avoid errors if duplicate drives are encountered
+
 ## 1.2.0 / 2021-04-23
 
 ### Changes